refactor: consistent types-ambient
