+++ packages/zoe/src/zoeService/zoeSeat.js
@@ -6,7 +6,7 @@ import { deeplyFulfilled } from '@endo/marshal';
 import { makePromiseKit } from '@endo/promise-kit';
 
 import { satisfiesWant } from '../contractFacet/offerSafety.js';
-import '../types.js';
+import '../types-ambient.js';
 import '../internal-types.js';
 import {
   declareOldZoeSeatAdminKind,
