+++ packages/zoe/typedoc.json
@@ -6,9 +6,9 @@
         "src/contractFacet/internal-types.js",
         "src/contractFacet/types-ambient.d.ts",
         "src/contractSupport/index.js",
-        "src/contractSupport/types.js",
+        "src/contractSupport/types-ambient.js",
         "src/types.js",
-        "src/zoeService/types.js",
+        "src/zoeService/types-ambient.js",
         "src/zoeService/utils.d.ts",
         "src/zoeService/zoe.js",
         "tools/internal-types.js",
